requirements.txt
aiogram==3.7.0
numpy
Pillow==10.4.0
scikit-learn==1.3.2